# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gym API",
        "description": "API pour la gestion de salle de sport",
        "contact": {
            "email": "admin@gym-gestion.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/subscription-types/{id}": {
            "get": {
                "tags": [
                    "Subscription Types"
                ],
                "summary": "Détails d'un type d'abonnement",
                "operationId": "b5c28e814b770d19c825f629a1816c14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du type de durée",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/subscription-types/new": {
            "post": {
                "tags": [
                    "Subscription Types"
                ],
                "summary": "Ajout de type d'abonnement",
                "operationId": "b0a86739f8556793188b94d6eab219a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "amount",
                                    "duration_type_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Premium"
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Abonnement premium"
                                    },
                                    "amount": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "49.99"
                                    },
                                    "duration_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "active_flag": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Enregistrement effectué"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/subscription-types/status/{id}": {
            "patch": {
                "tags": [
                    "Subscription Types"
                ],
                "summary": "Modifier le statut actif/inactif d'un type d'abonnement",
                "operationId": "07a0a609dff8eaf12eacea0069524ae2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du type d'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Modification effectuée"
                    },
                    "401": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/subscription-types/update/{id}": {
            "put": {
                "tags": [
                    "Subscription Types"
                ],
                "summary": "Mettre à jour les informations d'un type d'abonnement",
                "operationId": "8def9960ac133d2f2b15c7cc37bf4467",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du type d'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Premium"
                                    },
                                    "description": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Abonnement premium"
                                    },
                                    "amount": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "49.99"
                                    },
                                    "duration_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "active_flag": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mise à jour effectuée avec succès"
                    },
                    "401": {
                        "description": "Action non autorisée"
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
        "/api/subscription-types/delete/{id}": {
            "delete": {
                "tags": [
                    "Subscription Types"
                ],
                "summary": "Suppression d'un type d'abonnement",
                "operationId": "2fbbaffb2861a75aa6c6dd9291e1c190",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du type d'abonnement",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}